fix: Club Leader Response Model
