+++ data/club/src/commonMain/kotlin/com/b1nd/dodam/club/model/ClubMemberStudent.kt
@@ -1,0 +1,14 @@
+package com.b1nd.dodam.club.model
+
+
+data class ClubMemberStudent(
+    val id: Int,
+    val status: ClubState,
+    val permissions: ClubPermission,
+    val studentId: Int,
+    val name: String,
+    val grade: Int,
+    val room: Int,
+    val number: Int,
+    val profileImage: String?,
+)
