+++ network/club/src/commonMain/kotlin/com/b1nd/dodam/club/api/ClubService.kt
@@ -3,6 +3,7 @@ package com.b1nd.dodam.club.api
 import com.b1nd.dodam.club.datasource.ClubDataSource
 import com.b1nd.dodam.club.model.ClubJoinResponse
 import com.b1nd.dodam.club.model.ClubMemberResponse
+import com.b1nd.dodam.club.model.ClubMemberStudentResponse
 import com.b1nd.dodam.club.model.ClubResponse
 import com.b1nd.dodam.club.model.request.ClubJoinRequest
 import com.b1nd.dodam.club.model.request.ClubStateRequest
@@ -66,9 +67,9 @@ class ClubService(
         }.toImmutableList()
     }
 
-    override suspend fun getClubLeader(id: Int): ClubMemberResponse {
+    override suspend fun getClubLeader(id: Int): ClubMemberStudentResponse {
         return safeRequest {
-            client.get(DodamUrl.CLUB + "/$id/leader").body<Response<ClubMemberResponse>>()
+            client.get(DodamUrl.CLUB + "/$id/leader").body<Response<ClubMemberStudentResponse>>()
         }
     }
