+++ network/club/src/commonMain/kotlin/com/b1nd/dodam/club/datasource/ClubDataSource.kt
@@ -2,6 +2,7 @@ package com.b1nd.dodam.club.datasource
 
 import com.b1nd.dodam.club.model.ClubJoinResponse
 import com.b1nd.dodam.club.model.ClubMemberResponse
+import com.b1nd.dodam.club.model.ClubMemberStudentResponse
 import com.b1nd.dodam.club.model.ClubResponse
 import kotlinx.collections.immutable.ImmutableList
 
@@ -11,7 +12,7 @@ interface ClubDataSource {
     suspend fun deleteClubJoinRequests(id: Int)
     suspend fun getDetailClub(id: Int): ClubResponse
     suspend fun getClubJoinRequestReceived(): ImmutableList<ClubJoinResponse>
-    suspend fun getClubLeader(id: Int): ClubMemberResponse
+    suspend fun getClubLeader(id: Int): ClubMemberStudentResponse
     suspend fun getClubMember(id: Int): ClubMemberResponse
     suspend fun getClubList(): ImmutableList<ClubResponse>
     suspend fun getClubJoined(): ImmutableList<ClubResponse>
